+++ RELEASE_NOTES.md
@@ -1,13 +1,32 @@
-**2026.12.3**
+**2026.12.4**
+
+**Profiles**
+
+* Added a platform badge to show which platform your friend is currently using.
+
+**Friends Location Modal**
+
+* Added a platform badge to show which platform your friend is currently using.
+* Added Instance Owner Badge next to Instance type
+* Added Group Instance Owner next to Instance type
+
+**Dashboard Update**
+
+* Added **Edit Dashboard**. When right-clicking the dashboard item in the sidebar, a context menu will open.
+* Added **Favorite Worlds** dashboard element.
+* Added **Your Groups** dashboard element.
+* Added **Recently Visited** dashboard element.
+* Added **Popular Worlds** dashboard element.
+* Added **Very Active Worlds** dashboard element.
+* Added **My Recent Activity** dashboard element.
+* Added **Friends Recent Activity** dashboard element.
 
 **Changes**
 
-* Removed the old **Media Relay** design and replaced it with the **VRCN v2** design.
-* Reduced unused CSS classes for cleanup.
-* Removed dead CSS classes for cleanup.
+* Added a **Biography** title above the bio section in user profiles.
+* Added a **Show Full Bio** icon. When clicked, it displays the full bio so you do not have to scroll through long bios.
 
 **Fixes**
 
-* Fixed a bug in **Your Instances** that caused public instances to appear as your own, even though they were not actually yours.
-* #23 Fixed a bug that caused an infinite **"selecting database..."** loop when adding a VRCX database if the **Cancel** button was pressed.
-* #5 Fixed a bug that caused the music bar and time in **Custom Chatbox** to not refresh correctly.
+* Fixed an issue where profiles did not show separators.
+* Fixed an issue where profile bios did not display line breaks.